Apt: ACTIVATE _parseStream -- streams parse at resolve; handlers read the GUIAPT64 form
The activation pass, end to end:
* resolve64 now CALLS _parseStream on every action stream (case 1 @cmd+0x08), every
  clipActions record stream (case 3 inner loop), and every morph stream (case 8
  @cmd+0x10) -- serialized inline operands resolve into live pointers/values at
  movie-resolve time. The const ctx is the aptdata root (all GUIAPT64 offsets are
  root-relative, per libapt2; FLAG`d at the call sites) with a threaded value counter.
* The runtime handlers now read the parsed 8-ALIGNED GUIAPT64 operand form
  (_parseStream is the format authority): Push + DefineDictionary ({i64 count, table}
  +16), the PushString*/SetTarget/GotoLabel qword-pointer family (+8), GetUrl (2
  qwords), With (qword absolute end, +8), Try (24-byte record, sub-streams at +24),
  branch/GotoFrame/GotoFrame2/StoreRegister (align8 +4), and the DefineFunction/2
  records re-laid to the libapt2-verified 48-byte headers (args at 16-byte {reg,pad,
  name} / flat qword strides; body follows +48; the sig qwords 0x98765432/0x12345678
  are the runtime-patched constant-pool slots).
* PushWord/PushDWord/PushFloat/PushStringDictWord payloads read little-endian native
  (GUIAPT64 is the XB1 LE format; the console-BE forms are noted in comments).
* The console-form ResolveTranscode walker stays as the documented console-format
  sibling (dormant).

Gates clean (10 TUs) + full link clean. BOOT NOTE: dev HEAD is currently boot-blocked
UPSTREAM at loading stage 4 (BrnRootSoundModule.cpp:118 `lpSoundModuleInputBuffer !=
NULL`, from the TestBed-allocator sound wave -- fires identically WITHOUT this commit,
verified by a stash test; the Apt path is never reached). Apt-side boot verification
re-runs once the sound blocker clears.

+++ src/SDKs/EATech/include/Apt/AptScriptFunction1.h
@@ -59,15 +59,19 @@ class AptValue;
 // extra dword before the constant pool / bytecode), so it has its own record type;
 // this struct is the v1 form only.
 // ---------------------------------------------------------------------------
+// The GUIAPT64 48-byte serialized record (8-aligned; verified against the libapt2
+// converter's DefineFunction writer + the XB1 _parseStream walk). The two trailing
+// qwords carry serialized signatures (0x98765432 / 0x12345678) that the runtime
+// DefineFunction handler overwrites with the live constant-pool registers.
 struct AptScriptFunction1ByteCode
 {
     const char*  mpName;               // +0x00  the function's name (movie string; "" for anonymous)
-    int32_t      mnNumArguments;       // +0x04  declared parameter count
-    const char** mppArgumentNames;     // +0x08  array[mnNumArguments] of parameter-name strings
-    int32_t      mnByteCodeSize;       // +0x0C  length in bytes of the action bytecode
-    const char** mppConstantPool;      // +0x10  the constant-pool string table
-    int32_t      mnConstantPoolCount;  // +0x14  number of constant-pool entries
-    uint8_t      maByteCode[1];        // +0x18  the action bytecode (flexible; mnByteCodeSize bytes)
+    int64_t      mnNumArguments;       // +0x08  declared parameter count (u32 + pad)
+    const char** mppArgumentNames;     // +0x10  array[mnNumArguments] of qword name pointers
+    int64_t      mnByteCodeSize;       // +0x18  length in bytes of the action bytecode (u32 + pad)
+    const char** mppConstantPool;      // +0x20  sig1 slot -> the constant-pool table (patched in)
+    int64_t      mnConstantPoolCount;  // +0x28  sig2 slot -> number of entries (patched in)
+    uint8_t      maByteCode[1];        // +0x30  the action bytecode (flexible; mnByteCodeSize bytes)
 };
 
 class AptScriptFunction1 : public AptScriptFunctionBase
